Moves DetailCalloutListener implementation to CommCareActivity
